// src/core/schema.d.ts
export interface paths {
    "/": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Status
         * @description Return the status of the game server.
         *     This also includes a few global elements, such as announcements, server reset dates and leaderboards.
         */
        get: operations["get-status"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/my/agent": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Agent
         * @description Fetch your agent's details.
         */
        get: operations["get-my-agent"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/my/contracts": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * List Contracts
         * @description Return a paginated list of all your contracts.
         */
        get: operations["get-contracts"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/my/ships": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * List Ships
         * @description Return a paginated list of all of ships under your agent's ownership.
         */
        get: operations["get-my-ships"];
        put?: never;
        /**
         * Purchase Ship
         * @description Purchase a ship from a Shipyard. In order to use this function, a ship under your agent's ownership must be in a waypoint that has the `Shipyard` trait, and the Shipyard must sell the type of the desired ship.
         *
         *     Shipyards typically offer ship types, which are predefined templates of ships that have dedicated roles. A template comes with a preset of an engine, a reactor, and a frame. It may also include a few modules and mounts.
         */
        post: operations["purchase-ship"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        /**
         * @description The symbol of the faction.
         * @enum {string}
         */
        FactionSymbol: "COSMIC" | "VOID" | "GALACTIC" | "QUANTUM" | "DOMINION" | "ASTRO" | "CORSAIRS" | "OBSIDIAN" | "AEGIS" | "UNITED" | "SOLITARY" | "COBALT" | "OMEGA" | "ECHO" | "LORDS" | "CULT" | "ANCIENTS" | "SHADOW" | "ETHEREAL";
        /** @description Agent details. */
        Agent: {
            /** @description Account ID that is tied to this agent. Only included on your own agent. */
            accountId?: string;
            /** @description Symbol of the agent. */
            symbol: string;
            /** @description The headquarters of the agent. */
            headquarters: string;
            /**
             * Format: int64
             * @description The number of credits the agent has available. Credits can be negative if funds have been overdrawn.
             */
            credits: number;
            /** @description The faction the agent started with. */
            startingFaction: string;
            /** @description How many ships are owned by the agent. */
            shipCount: number;
        };
// ...
        Contract: {
            /** @description ID of the contract. */
            id: string;
            /** @description The symbol of the faction that this contract is for. */
            factionSymbol: string;
            /**
             * @description Type of contract.
             * @enum {string}
             */
            type: "PROCUREMENT" | "TRANSPORT" | "SHUTTLE";
            terms: components["schemas"]["ContractTerms"];
            /**
             * @description Whether the contract has been accepted by the agent
             * @default false
             */
            accepted: boolean;
            /**
             * @description Whether the contract has been fulfilled
             * @default false
             */
            fulfilled: boolean;
            /**
             * Format: date-time
             * @deprecated
             * @description Deprecated in favor of deadlineToAccept
             */
            expiration: string;
            /**
             * Format: date-time
             * @description The time at which the contract is no longer available to be accepted
             */
            deadlineToAccept?: string;
        };
// ...
        Faction: {
            symbol: components["schemas"]["FactionSymbol"];
            /** @description Name of the faction. */
            name: string;
            /** @description Description of the faction. */
            description: string;
            /** @description The waypoint in which the faction's HQ is located in. */
            headquarters?: string;
            /** @description List of traits that define this faction. */
            traits: components["schemas"]["FactionTrait"][];
            /** @description Whether or not the faction is currently recruiting new agents. */
            isRecruiting: boolean;
        };
// ...
        ShipCargo: {
            /** @description The max number of items that can be stored in the cargo hold. */
            capacity: number;
            /** @description The number of items currently stored in the cargo hold. */
            units: number;
            /** @description The items currently in the cargo hold. */
            inventory: components["schemas"]["ShipCargoItem"][];
        };
// ...
        Ship: {
            /** @description The globally unique identifier of the ship in the following format: `[AGENT_SYMBOL]-[HEX_ID]` */
            symbol: string;
            registration: components["schemas"]["ShipRegistration"];
            nav: components["schemas"]["ShipNav"];
            crew: components["schemas"]["ShipCrew"];
            frame: components["schemas"]["ShipFrame"];
            reactor: components["schemas"]["ShipReactor"];
            engine: components["schemas"]["ShipEngine"];
            cooldown: components["schemas"]["Cooldown"];
            /** @description Modules installed in this ship. */
            modules: components["schemas"]["ShipModule"][];
            /** @description Mounts installed in this ship. */
            mounts: components["schemas"]["ShipMount"][];
            cargo: components["schemas"]["ShipCargo"];
            fuel: components["schemas"]["ShipFuel"];
        };
// ...
        System: {
            /** @description The symbol of the system. */
            symbol: string;
            /** @description The symbol of the sector. */
            sectorSymbol: string;
            /** @description The constellation that the system is part of. */
            constellation?: string;
            /** @description The name of the system. */
            name?: string;
            type: components["schemas"]["SystemType"];
            /** @description Relative position of the system in the sector in the x axis. */
            x: number;
            /** @description Relative position of the system in the sector in the y axis. */
            y: number;
            /** @description Waypoints in this system. */
            waypoints: components["schemas"]["SystemWaypoint"][];
            /** @description Factions that control this system. */
            factions: components["schemas"]["SystemFaction"][];
        };
// ...
        Waypoint: {
            symbol: components["schemas"]["WaypointSymbol"];
            type: components["schemas"]["WaypointType"];
            systemSymbol: components["schemas"]["SystemSymbol"];
            /** @description Relative position of the waypoint on the system's x axis. This is not an absolute position in the universe. */
            x: number;
            /** @description Relative position of the waypoint on the system's y axis. This is not an absolute position in the universe. */
            y: number;
            /** @description Waypoints that orbit this waypoint. */
            orbitals: components["schemas"]["WaypointOrbital"][];
            /** @description The symbol of the parent waypoint, if this waypoint is in orbit around another waypoint. Otherwise this value is undefined. */
            orbits?: string;
            faction?: components["schemas"]["WaypointFaction"];
            /** @description The traits of the waypoint. */
            traits: components["schemas"]["WaypointTrait"][];
            /** @description The modifiers of the waypoint. */
            modifiers?: components["schemas"]["WaypointModifier"][];
            chart?: components["schemas"]["Chart"];
            /** @description True if the waypoint is under construction. */
            isUnderConstruction: boolean;
        };
// ...
        Market: {
            /** @description The symbol of the market. The symbol is the same as the waypoint where the market is located. */
            symbol: string;
            /** @description The list of goods that are exported from this market. */
            exports: components["schemas"]["TradeGood"][];
            /** @description The list of goods that are sought as imports in this market. */
            imports: components["schemas"]["TradeGood"][];
            /** @description The list of goods that are bought and sold between agents at this market. */
            exchange: components["schemas"]["TradeGood"][];
            /** @description The list of recent transactions at this market. Visible only when a ship is present at the market. */
            transactions?: components["schemas"]["MarketTransaction"][];
            /** @description The list of goods that are traded at this market. Visible only when a ship is present at the market. */
            tradeGoods?: components["schemas"]["MarketTradeGood"][];
        };
// ...
        ShipyardTransaction: {
            waypointSymbol: components["schemas"]["WaypointSymbol"];
            /**
             * @deprecated
             * @description The symbol of the ship type (e.g. SHIP_MINING_DRONE) that was the subject of the transaction. Contrary to what the name implies, this is NOT the symbol of the ship that was purchased.
             */
            shipSymbol: string;
            /** @description The symbol of the ship type (e.g. SHIP_MINING_DRONE) that was the subject of the transaction. */
            shipType: string;
            /** @description The price of the transaction. */
            price: number;
            /** @description The symbol of the agent that made the transaction. */
            agentSymbol: string;
            /**
             * Format: date-time
             * @description The timestamp of the transaction.
             */
            timestamp: string;
        };
// ...
        Shipyard: {
            /** @description The symbol of the shipyard. The symbol is the same as the waypoint where the shipyard is located. */
            symbol: string;
            /** @description The list of ship types available for purchase at this shipyard. */
            shipTypes: {
                type: components["schemas"]["ShipType"];
            }[];
            /** @description The list of recent transactions at this shipyard. */
            transactions?: components["schemas"]["ShipyardTransaction"][];
            /** @description The ships that are currently available for purchase at the shipyard. */
            ships?: components["schemas"]["ShipyardShip"][];
            /** @description The fee to modify a ship at this shipyard. This includes installing or removing modules and mounts on a ship. In the case of mounts, the fee is a flat rate per mount. In the case of modules, the fee is per slot the module occupies. */
            modificationsFee: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    "get-status": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
// ...
    };
// ...
}